// src/api/model/device.ts
/**
 * LinkIOT Dashboard
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Device { 
    name: string;
    deviceId: string;
    secret: string;
    description: string;
    state: number;
    status: Boolean;
    lastLoginTime: string;
    image: string;
}